Fixed bugs and added more and different sprites
(kind of) Fixed bug where segfault happens after calling set_tile_color.
If cout is called afterwards it does not segfaults ????
Fixed bug where agents could not selected
Added debug print outs and remove ones that spam too much. Should add
some nicer solution.
Fixed bug where hex aplha value was garbage.
Made nicer print of sdl_color

+++ src/sdl_helper.hpp
@@ -31,7 +31,7 @@ static std::ostream& operator<<(std::ostream& ost, const SDL_Rect& r)
 static std::ostream& operator<<(std::ostream& ost, const SDL_Color& c)
 {
 	ost << '[';
-	ost << c.r << ", " << c.g << ", " << c.b << ", " << c.a;
+	ost << (int)c.r << ", " << (int)c.g << ", " << (int)c.b << ", " << (int)c.a;
 	ost << ']';
 	return ost;
 }
